refactor(schema): deprecate 'PreviewerConfig'
