+++ lua/fzfx/general.lua
@@ -8,53 +8,10 @@ local env = require("fzfx.env")
 local path = require("fzfx.path")
 local ProviderTypeEnum = require("fzfx.schema").ProviderTypeEnum
 local PreviewerTypeEnum = require("fzfx.schema").PreviewerTypeEnum
-local clazz = require("fzfx.clazz")
-local PreviewerConfig = require("fzfx.schema").PreviewerConfig
+local schema = require("fzfx.schema")
 local conf = require("fzfx.config")
 
 local DEFAULT_PIPELINE = "default"
-
--- config class detect {
-
---- @param cfg Configs?
---- @return boolean
-local function is_command_config(cfg)
-    return type(cfg) == "table"
-        and type(cfg.name) == "string"
-        and string.len(cfg.name) > 0
-        and type(cfg.feed) == "string"
-        and string.len(cfg.feed) > 0
-        and type(cfg.opts) == "table"
-end
-
---- @param cfg Configs?
---- @return boolean
-local function is_provider_config(cfg)
-    return type(cfg) == "table"
-        and type(cfg.key) == "string"
-        and string.len(cfg.key) > 0
-        and (
-            (type(cfg.provider) == "string" and string.len(cfg.provider) > 0)
-            or (type(cfg.provider) == "table" and #cfg.provider > 0)
-            or type(cfg.provider) == "function"
-        )
-end
-
---- @param cfg Configs?
---- @return boolean
-local function is_previewer_config(cfg)
-    return type(cfg) == "table"
-        and type(cfg.previewer) == "function"
-        and (
-            cfg.previewer_type == nil
-            or (
-                type(cfg.previewer_type) == "string"
-                and string.len(cfg.previewer_type) > 0
-            )
-        )
-end
-
--- config class detect }
 
 -- provider switch {
 
@@ -101,21 +58,21 @@ local ProviderSwitch = {}
 --- @return ProviderSwitch
 function ProviderSwitch:new(name, pipeline, provider_configs)
     local provider_configs_map = {}
-    if is_provider_config(provider_configs) then
+    if schema.is_provider_config(provider_configs) then
         provider_configs.provider_type =
-            get_provider_type_or_default(provider_configs)
+            schema.get_provider_type_or_default(provider_configs)
         provider_configs_map[DEFAULT_PIPELINE] = provider_configs
     else
         for provider_name, provider_opts in pairs(provider_configs) do
             log.ensure(
-                is_provider_config(provider_opts),
+                schema.is_provider_config(provider_opts),
                 "error! %s (%s) is not a valid provider! %s",
                 vim.inspect(provider_name),
                 vim.inspect(name),
                 vim.inspect(provider_opts)
             )
             provider_opts.provider_type =
-                get_provider_type_or_default(provider_opts)
+                schema.get_provider_type_or_default(provider_opts)
             provider_configs_map[provider_name] = provider_opts
         end
     end
@@ -379,15 +336,15 @@ local PreviewerSwitch = {}
 function PreviewerSwitch:new(name, pipeline, previewer_configs)
     local previewers_map = {}
     local previewer_types_map = {}
-    if is_previewer_config(previewer_configs) then
+    if schema.is_previewer_config(previewer_configs) then
         previewers_map[DEFAULT_PIPELINE] = previewer_configs.previewer
         previewer_types_map[DEFAULT_PIPELINE] =
-            get_previewer_type_or_default(previewer_configs)
+            schema.get_previewer_type_or_default(previewer_configs)
     else
         for previewer_name, previewer_opts in pairs(previewer_configs) do
             previewers_map[previewer_name] = previewer_opts.previewer
             previewer_types_map[previewer_name] =
-                get_previewer_type_or_default(previewer_opts)
+                schema.get_previewer_type_or_default(previewer_opts)
         end
     end
 
@@ -600,7 +557,7 @@ end
 --- @return HeaderSwitch
 function HeaderSwitch:new(provider_configs, interaction_configs)
     local headers_map = {}
-    if is_provider_config(provider_configs) then
+    if schema.is_provider_config(provider_configs) then
         headers_map[DEFAULT_PIPELINE] = make_help_doc(interaction_configs, {})
     else
         log.debug(
@@ -645,7 +602,7 @@ end
 local function get_pipeline_size(pipeline_configs)
     local n = 0
     if type(pipeline_configs) == "table" then
-        if is_provider_config(pipeline_configs.providers) then
+        if schema.is_provider_config(pipeline_configs.providers) then
             return 1
         end
         for _, _ in pairs(pipeline_configs.providers) do
@@ -677,7 +634,7 @@ local function general(name, query, bang, pipeline_configs, default_pipeline)
     if default_pipeline == nil then
         local pipeline = nil
         local provider_opts = nil
-        if is_provider_config(pipeline_configs.providers) then
+        if schema.is_provider_config(pipeline_configs.providers) then
             log.debug(
                 "|fzfx.general - general| providers is single config: %s",
                 vim.inspect(pipeline_configs.providers)
@@ -931,7 +888,7 @@ local function setup(name, pipeline_configs)
     --     vim.inspect(pipeline_configs)
     -- )
     -- User commands
-    if is_command_config(pipeline_configs.commands) then
+    if schema.is_command_config(pipeline_configs.commands) then
         vim.api.nvim_create_user_command(
             pipeline_configs.commands.name,
             function(opts)
@@ -972,11 +929,6 @@ end
 
 local M = {
     setup = setup,
-    is_command_config = is_command_config,
-    is_provider_config = is_provider_config,
-    is_previewer_config = is_previewer_config,
-    get_provider_type_or_default = get_provider_type_or_default,
-    get_previewer_type_or_default = get_previewer_type_or_default,
     make_cache_filename = make_cache_filename,
     ProviderSwitch = ProviderSwitch,
     PreviewerSwitch = PreviewerSwitch,
